feat: added shop_id

+++ app/Http/Controllers/ExpenseController.php
@@ -10,7 +10,7 @@ class ExpenseController extends Controller
     {
         return response()->json([
             'success' => true,
-            'data' => Expense::with('user')->orderBy('date','desc')->get(),
+            'data' => Expense::where('shop_id', Auth::user()->shop_id)->with('user')->orderBy('date','desc')->get(),
             'message' => 'Expenses retrieved successfully'
         ]);
     }
@@ -25,6 +25,7 @@ class ExpenseController extends Controller
         ]);
 
         $data['recorded_by'] = $r->user()->id;
+        $data['shop_id'] = Auth::user()->shop_id;
         return response()->json([
             'success' => true,
             'data' => Expense::create($data),
@@ -36,14 +37,14 @@ class ExpenseController extends Controller
     {
         return response()->json([
             'success' => true,
-            'data' => Expense::with('user')->findOrFail($id),
+            'data' => Expense::where('shop_id', Auth::user()->shop_id)->with('user')->findOrFail($id),
             'message' => 'Expense retrieved successfully'
         ]);
     }
 
     public function update(Request $r, $id)
     {
-        $expense = Expense::findOrFail($id);
+        $expense = Expense::where('shop_id', Auth::user()->shop_id)->findOrFail($id);
         $expense->update($r->all());
         return response()->json([
             'success' => true,
@@ -54,7 +55,7 @@ class ExpenseController extends Controller
 
     public function destroy($id)
     {
-        Expense::findOrFail($id)->delete();
+        Expense::where('shop_id', Auth::user()->shop_id)->findOrFail($id)->delete();
         return response()->json([
             'success' => true,
             'message' => 'Expense deleted successfully'
